feat: macro shortcuts
