+++ module/base/sheet/actor/methods/enhancement-methods.mjs
@@ -12,7 +12,7 @@ import { EnhancementRepository } from "../../../../repository/enhancement-reposi
 import { ActorUpdater } from "../../../updater/actor-updater.mjs";
 import { ActiveEffectsUtils } from "../../../../core/effect/active-effects.mjs";
 import { ActiveEffectsOriginTypes } from "../../../../enums/active-effects-enums.mjs";
-import { ActorUtils } from "../../../../utils/actor-utils.mjs";
+import { ActorUtils } from "../../../../core/actor/actor-utils.mjs";
 
 export function updateEnhancementLevelsOptions(enhancementId, selects) {
     const enhancementLevels = EnhancementRepository._getEnhancementEffectsByEnhancementId(enhancementId);
