+++ module/core/actor/actor-equipment.mjs
