+++ module/core/rolls/initiative-roll.mjs
@@ -1,4 +1,4 @@
-import { ActorUtils } from "../../utils/actor-utils.mjs";
+import { ActorUtils } from "../actor/actor-utils.mjs";
 import { CoreRollMethods } from "./core-roll-methods.mjs";
 
 export class RollInitiative {
