+++ styles/system-styles.css
@@ -26,6 +26,7 @@ body {
     --on-tertiary-alternate-color: var(--primary-color);
 
     --color-gray: rgba(80, 80, 80, 1);
+    --color-gray-50: rgba(80, 80, 80, 0.5);
     --color-gray-alpha: rgba(80, 80, 80, 0.8);
     --color-black-alpha: rgba(0, 0, 0, 0.8);
     --color-black-alpha-50: rgba(0, 0, 0, 0.5);
@@ -122,9 +123,22 @@ section {
 
 /* #region EDIT FOUNDRY STYLE */
 
-.compendium.directory .directory-list .directory-item {
+.directory-list .directory-item {
     padding-block: 2px;
     box-shadow: 0px 1px 6px black;
+}
+
+.compendium.directory .directory-item.document {
+    color: var(--primary-color);
+    background-color: var(--color-gray-50);
+}
+
+.compendium.directory .directory-item.document img {
+    margin-inline: 6px;
+}
+
+.compendium.directory .directory-item.document h4 {
+    margin: 0px;
 }
 
 img#logo {
