+++ templates/items/equipment-bag-item.hbs
