+++ templates/items/equipment-equipped-item.hbs
